pyproject.toml
[project]
name = "updatebot"
version = "1.0.0"
description = "Auto-update bot for ZJU CSSU website notifications"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
updatebot = "updatebot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
